requirements.txt
Flask==3.0.0
Requests==2.31.0
py-clob-client==0.34.6
python-dotenv==1.0.1
gunicorn==21.2.0
